fix(trusty-code): reject task.run project/session_id binding mismatch
Code-critic HIGH finding on PR #3189: when session_id reused an existing
session AND project resolved to a different root than that session's
persisted Session.binding, the run would execute against the new project
while session.status/session.list kept reporting the original binding
forever (SessionRegistry has no binding-update path). Reject the
combination instead: a per-call project may only restate an existing
session's own binding root, never silently redirect it, mapping a
mismatch to -32003 invalid_argument (400 over REST) naming both roots.

Also splits task/protocol.rs's inline test module into protocol_tests.rs
(the crate's established _tests.rs sibling-file convention) to keep the
production file under the 500-SLOC cap after the added validation.
