+++ crates/trusty-code/src/serve/rest/tasks.rs
@@ -323,4 +323,37 @@ mod tests {
         let v = body_json(resp).await;
         assert_eq!(v["error"]["code"], -32003);
     }
+
+    /// A `session_id` reusing an existing session, paired with a `project`
+    /// naming a DIFFERENT root than that session's own persisted binding,
+    /// must be rejected with `400`/`-32003 invalid_argument` (#3178,
+    /// code-critic HIGH finding, PR #3189) — the REST-level half of
+    /// `task::protocol::tests::task_run_session_id_with_mismatched_project_is_rejected`.
+    #[tokio::test]
+    async fn run_task_session_id_with_mismatched_project_returns_400() {
+        let (app, sessions, _boot_project) = app_and_registry();
+        let session_project = tempfile::tempdir().expect("session project tempdir");
+        let other_project = tempfile::tempdir().expect("other project tempdir");
+        let session_binding = ProjectBinding::resolve(Some(session_project.path().to_path_buf()))
+            .expect("tempdir must bind");
+        let existing = sessions.create("say hi".to_string(), None, session_binding);
+
+        let body = json!({
+            "task_description": "say hi",
+            "session_id": existing.id,
+            "project": other_project.path().to_string_lossy(),
+        })
+        .to_string();
+        let resp = post(&app, "/tasks", &body).await;
+        assert_eq!(resp.status(), StatusCode::BAD_REQUEST);
+        let v = body_json(resp).await;
+        assert_eq!(v["error"]["code"], -32003);
+        assert!(
+            v["error"]["message"]
+                .as_str()
+                .unwrap_or_default()
+                .contains("does not match session"),
+            "error message must name the mismatch: {v}"
+        );
+    }
 }
